requirements.txt
asgiref==3.4.1
asyncio==3.4.3
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
cryptography==37.0.4
Django==3.2.15
django-rest-knox==4.2.0
djangorestframework==3.13.1
entrypoints==0.4
idna==3.3
jeepney==0.8.0
keyring==18.0.1
mysqlclient==2.1.1
numpy==1.22.4
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2021.3
PyYAML==5.3.1
requests==2.27.1
SecretStorage==3.3.3
sqlparse==0.4.2
urllib3==1.26.11
